requirements.txt
PyYAML
gunicorn==19.9.0
Flask==1.0.2
Jinja2==2.10
psycopg2==2.7.6.1
Werkzeug==0.14.1
gunicorn==19.9.0
requests==2.20.1
flask_heroku==0.1.9
nose==1.3.7
nosexcover==1.0.11
pylint==2.2.2
sentry_sdk==0.7.6
kafka-python==1.4.6
confluent-kafka==1.1.0